chromewhip-master/requirements.txt
aiohttp==3.6.2
pytest-asyncio==0.10.0
websockets==7.0
beautifulsoup4==4.7.1
lxml==4.6.2
pyyaml==5.1
Pillow==7.1.0
